requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-docx>=1.1.0
pdfplumber>=0.10.0
openpyxl>=3.1.2
requests>=2.31.0
anthropic>=0.28.0
python-dotenv>=1.0.1
pydantic>=2.7.0
pydantic-settings>=2.2.0
aiofiles>=23.2.1
beautifulsoup4>=4.12.0
lxml>=5.0.0
